pyproject.toml
[project]
name = "dinzbot"
version = "0.1.0"
description = ""
authors = [
    {name = "DinzGIT",email = "ygg33333@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite (>=0.22.1,<0.23.0)",
    "discord-py (>=2.7.1,<3.0.0)",
    "pytz (>=2026.1.post1,<2027.0)",
    "openai (>=2.29.0,<3.0.0)",
    "tiktoken (>=0.12.0,<0.13.0)",
    "pillow (>=12.1.1,<13.0.0)",
    "python-dotenv (>=1.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
